// src/app/interfaces/common/management.interface.ts
export interface Management {
  _id?: string;
  name?: string;
  slug?: string;
  priority?: number;
  image?: string;
  mobileImage?: string;
  bannerType?: string;
  url?: string;
  email?: string;
  managementType?: string;
  phone?: string;
  designation?: string;
  createdAt?: Date;
  updatedAt?: Date;
}